chore(app): Move the ErrorBoundary version of FreshdeskWidget to a better location
