+++ packages/openneuro-app/src/scripts/errors/freshdesk-widget.jsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import { useCookies } from 'react-cookie'
-import { getProfile } from '../../authentication/profile'
-import { config } from '../../config'
+import { getProfile } from '../authentication/profile'
+import { config } from '../config'
 
 const buildCustomQuery = (customText, prepopulatedFields) => {
   const customizerQueries = [
